feat: Implement user session persistence
This change introduces logic to persist user sessions using local storage. Upon application initialization, it attempts to restore a previously saved user session. If a session is found, the user is logged in and directed to the dashboard. If not, or if restoration fails, the user is directed to the landing page. User sessions are now saved to local storage whenever a user logs in or their information is updated.

This also includes:
- Moving Tailwind CSS script to the end of head in `index.html` for better rendering performance.
- Updating the environment variable retrieval in `firebaseService.ts` to prioritize runtime variables injected via `env-config.js`.
- Adding a link to the YouTube video on the thumbnail in `ChannelDetailView.tsx` for direct access.
- Implementing cache checks for video details in `VideoDetailView.tsx` to improve performance by avoiding redundant API calls.